Bool optimizer is stateless
