+++ ecmascript/minifier/src/compress/mod.rs
@@ -42,6 +42,7 @@ mod drop_console;
 mod hoist_decls;
 mod optimize;
 mod pure;
+mod util;
 
 pub(crate) fn compressor<'a>(
     cm: Lrc<SourceMap>,
