+++ ecmascript/minifier/src/compress/optimize/bools.rs
@@ -1,5 +1,6 @@
 use super::Optimizer;
 use crate::compress::optimize::{is_pure_undefined, Ctx};
+use crate::compress::util::negate_cost;
 use crate::debug::dump;
 use crate::util::make_bool;
 use swc_atoms::js_word;
@@ -8,7 +9,6 @@ use swc_ecma_ast::*;
 use swc_ecma_transforms_base::ext::MapWithMut;
 use swc_ecma_utils::ident::IdentLike;
 use swc_ecma_utils::Type;
-use swc_ecma_utils::Value;
 use swc_ecma_utils::Value::Known;
 use swc_ecma_utils::Value::Unknown;
 use swc_ecma_utils::{undefined, ExprExt};
@@ -321,158 +321,6 @@ impl Optimizer<'_> {
         }
     }
 
-    /// This method converts `!1` to `0`.
-    pub(super) fn optimize_expr_in_bool_ctx(&mut self, n: &mut Expr) {
-        if !self.options.bools {
-            return;
-        }
-
-        match n {
-            Expr::Bin(BinExpr {
-                op: op!("&&") | op!("||"),
-                left,
-                right,
-                ..
-            }) => {
-                // Regardless if it's truthy or falsy, we can optimize it because it will be
-                // casted as bool anyway.
-                self.optimize_expr_in_bool_ctx(&mut **left);
-                self.optimize_expr_in_bool_ctx(&mut **right);
-                return;
-            }
-
-            Expr::Seq(e) => {
-                if let Some(last) = e.exprs.last_mut() {
-                    self.optimize_expr_in_bool_ctx(&mut **last);
-                }
-            }
-
-            _ => {}
-        }
-
-        match n {
-            Expr::Unary(UnaryExpr {
-                span,
-                op: op!("!"),
-                arg,
-            }) => match &mut **arg {
-                Expr::Lit(Lit::Num(Number { value, .. })) => {
-                    log::debug!("Optimizing: number => number (in bool context)");
-
-                    self.changed = true;
-                    *n = Expr::Lit(Lit::Num(Number {
-                        span: *span,
-                        value: if *value == 0.0 { 1.0 } else { 0.0 },
-                    }))
-                }
-
-                Expr::Unary(UnaryExpr {
-                    op: op!("!"), arg, ..
-                }) => {
-                    log::debug!("bools: !!expr => expr (in bool ctx)");
-                    self.changed = true;
-                    *n = *arg.take();
-                    return;
-                }
-                _ => {}
-            },
-
-            Expr::Unary(UnaryExpr {
-                span,
-                op: op!("typeof"),
-                arg,
-            }) => {
-                log::debug!("Optimizing: typeof => true (in bool context)");
-                self.changed = true;
-
-                match &**arg {
-                    Expr::Ident(..) => {
-                        *n = Expr::Lit(Lit::Num(Number {
-                            span: *span,
-                            value: 1.0,
-                        }))
-                    }
-                    _ => {
-                        // Return value of typeof is always truthy
-                        let true_expr = Box::new(Expr::Lit(Lit::Num(Number {
-                            span: *span,
-                            value: 1.0,
-                        })));
-                        *n = Expr::Seq(SeqExpr {
-                            span: *span,
-                            exprs: vec![arg.take(), true_expr],
-                        })
-                    }
-                }
-            }
-
-            Expr::Lit(Lit::Str(s)) => {
-                log::debug!("Converting string as boolean expressions");
-                self.changed = true;
-                *n = Expr::Lit(Lit::Num(Number {
-                    span: s.span,
-                    value: if s.value.is_empty() { 0.0 } else { 1.0 },
-                }));
-            }
-
-            Expr::Lit(Lit::Num(num)) => {
-                if num.value == 1.0 || num.value == 0.0 {
-                    return;
-                }
-                if self.options.bools {
-                    log::debug!("booleans: Converting number as boolean expressions");
-                    self.changed = true;
-                    *n = Expr::Lit(Lit::Num(Number {
-                        span: num.span,
-                        value: if num.value == 0.0 { 0.0 } else { 1.0 },
-                    }));
-                }
-            }
-
-            Expr::Bin(BinExpr {
-                op: op!("??"),
-                left,
-                right,
-                ..
-            }) => {
-                // Optimize if (a ?? false); as if (a);
-                if let Value::Known(false) = right.as_pure_bool() {
-                    log::debug!(
-                        "Dropping right operand of `??` as it's always false (in bool context)"
-                    );
-                    self.changed = true;
-                    *n = *left.take();
-                }
-            }
-
-            Expr::Bin(BinExpr {
-                op: op!("||"),
-                left,
-                right,
-                ..
-            }) => {
-                // `a || false` => `a` (as it will be casted to boolean anyway)
-
-                if let Known(false) = right.as_pure_bool() {
-                    log::debug!("bools: `expr || false` => `expr` (in bool context)");
-                    self.changed = true;
-                    *n = *left.take();
-                    return;
-                }
-            }
-
-            _ => {
-                let span = n.span();
-                let v = n.as_pure_bool();
-                if let Known(v) = v {
-                    log::debug!("Optimizing expr as {} (in bool context)", v);
-                    *n = make_bool(span, v);
-                    return;
-                }
-            }
-        }
-    }
-
     pub(super) fn compress_if_stmt_as_expr(&mut self, s: &mut Stmt) {
         if !self.options.bools {
             return;
@@ -560,174 +408,6 @@ impl Optimizer<'_> {
             };
         }
     }
-}
-
-pub(crate) fn is_ok_to_negate_for_cond(e: &Expr) -> bool {
-    match e {
-        Expr::Update(..) => false,
-        _ => true,
-    }
-}
-
-pub(crate) fn is_ok_to_negate_rhs(rhs: &Expr) -> bool {
-    match rhs {
-        Expr::Member(..) => true,
-        Expr::Bin(BinExpr {
-            op: op!("===") | op!("!==") | op!("==") | op!("!="),
-            ..
-        }) => true,
-
-        Expr::Call(..) | Expr::New(..) => false,
-
-        Expr::Update(..) => false,
-
-        Expr::Bin(BinExpr {
-            op: op!("&&") | op!("||"),
-            left,
-            right,
-            ..
-        }) => is_ok_to_negate_rhs(&left) && is_ok_to_negate_rhs(&right),
-
-        Expr::Bin(BinExpr { left, right, .. }) => {
-            is_ok_to_negate_rhs(&left) && is_ok_to_negate_rhs(&right)
-        }
-
-        Expr::Assign(e) => is_ok_to_negate_rhs(&e.right),
-
-        Expr::Unary(UnaryExpr {
-            op: op!("!") | op!("delete"),
-            ..
-        }) => true,
-
-        Expr::Seq(e) => {
-            if let Some(last) = e.exprs.last() {
-                is_ok_to_negate_rhs(&last)
-            } else {
-                true
-            }
-        }
-
-        Expr::Cond(e) => is_ok_to_negate_rhs(&e.cons) && is_ok_to_negate_rhs(&e.alt),
-
-        _ => {
-            if !rhs.may_have_side_effects() {
-                return true;
-            }
-
-            if cfg!(feature = "debug") {
-                log::warn!("unimplemented: is_ok_to_negate_rhs: `{}`", dump(&*rhs));
-            }
-
-            false
-        }
-    }
-}
-
-/// A negative value means that it's efficient to negate the expression.
-pub(crate) fn negate_cost(e: &Expr, in_bool_ctx: bool, is_ret_val_ignored: bool) -> Option<isize> {
-    fn cost(
-        e: &Expr,
-        in_bool_ctx: bool,
-        bin_op: Option<BinaryOp>,
-        is_ret_val_ignored: bool,
-    ) -> isize {
-        match e {
-            Expr::Unary(UnaryExpr {
-                op: op!("!"), arg, ..
-            }) => {
-                // TODO: Check if this argument is actually start of line.
-                match &**arg {
-                    Expr::Call(CallExpr {
-                        callee: ExprOrSuper::Expr(callee),
-                        ..
-                    }) => match &**callee {
-                        Expr::Fn(..) => return 0,
-                        _ => {}
-                    },
-                    _ => {}
-                }
-
-                if in_bool_ctx {
-                    let c = -cost(arg, true, None, is_ret_val_ignored);
-                    return c.min(-1);
-                }
-
-                match &**arg {
-                    Expr::Unary(UnaryExpr { op: op!("!"), .. }) => -1,
-
-                    _ => 1,
-                }
-            }
-            Expr::Bin(BinExpr {
-                op: op!("===") | op!("!==") | op!("==") | op!("!="),
-                ..
-            }) => 0,
-
-            Expr::Bin(BinExpr {
-                op: op @ op!("||") | op @ op!("&&"),
-                left,
-                right,
-                ..
-            }) => {
-                let l_cost = cost(&left, in_bool_ctx, Some(*op), false);
-
-                if !is_ret_val_ignored && !is_ok_to_negate_rhs(&right) {
-                    return l_cost + 3;
-                }
-                l_cost + cost(&right, in_bool_ctx, Some(*op), is_ret_val_ignored)
-            }
-
-            Expr::Cond(CondExpr { cons, alt, .. })
-                if is_ok_to_negate_for_cond(&cons) && is_ok_to_negate_for_cond(&alt) =>
-            {
-                cost(&cons, in_bool_ctx, bin_op, is_ret_val_ignored)
-                    + cost(&alt, in_bool_ctx, bin_op, is_ret_val_ignored)
-            }
-
-            Expr::Cond(..)
-            | Expr::Update(..)
-            | Expr::Bin(BinExpr {
-                op: op!("in") | op!("instanceof"),
-                ..
-            }) => 3,
-
-            Expr::Assign(..) => {
-                if is_ret_val_ignored {
-                    0
-                } else {
-                    3
-                }
-            }
-
-            Expr::Seq(e) => {
-                if let Some(last) = e.exprs.last() {
-                    return cost(&last, in_bool_ctx, bin_op, is_ret_val_ignored);
-                }
-
-                if is_ret_val_ignored {
-                    0
-                } else {
-                    1
-                }
-            }
-
-            _ => {
-                if is_ret_val_ignored {
-                    0
-                } else {
-                    1
-                }
-            }
-        }
-    }
-
-    let cost = cost(e, in_bool_ctx, None, is_ret_val_ignored);
-
-    if cfg!(feature = "debug") {
-        log::trace!("negation cost of `{}` = {}", dump(&*e), cost);
-    }
-
-    Some(cost)
 }
 
 #[cfg(test)]
