+++ ecmascript/minifier/src/compress/optimize/mod.rs
@@ -1634,8 +1634,6 @@ impl VisitMut for Optimizer<'_> {
         n.visit_mut_children_with(self);
 
         self.negate_cond_expr(n);
-
-        self.optimize_expr_in_bool_ctx(&mut n.test);
     }
 
     fn visit_mut_decl(&mut self, decl: &mut Decl) {
@@ -1750,7 +1748,6 @@ impl VisitMut for Optimizer<'_> {
         self.compress_cond_with_logical_as_logical(e);
 
         self.compress_negated_bin_eq(e);
-        self.handle_negated_seq(e);
         self.compress_array_join(e);
 
         self.remove_useless_pipes(e);
@@ -1922,10 +1919,6 @@ impl VisitMut for Optimizer<'_> {
 
         self.with_ctx(ctx).drop_if_break(s);
 
-        if let Some(test) = &mut s.test {
-            self.with_ctx(ctx).optimize_expr_in_bool_ctx(test);
-        }
-
         match &mut *s.body {
             Stmt::Block(body) => {
                 self.negate_if_terminate(&mut body.stmts, false, true);
@@ -2000,8 +1993,6 @@ impl VisitMut for Optimizer<'_> {
         n.alt.visit_mut_with(&mut *self.with_ctx(ctx));
 
         self.negate_if_stmt(n);
-
-        self.optimize_expr_in_bool_ctx(&mut n.test);
 
         self.merge_nested_if(n);
 
@@ -2396,9 +2387,7 @@ impl VisitMut for Optimizer<'_> {
 
         n.visit_mut_children_with(&mut *self.with_ctx(ctx));
 
-        if n.op == op!("!") {
-            self.with_ctx(ctx).optimize_expr_in_bool_ctx(&mut n.arg);
-        } else if n.op == op!(unary, "+") || n.op == op!(unary, "-") {
+        if n.op == op!(unary, "+") || n.op == op!(unary, "-") {
             self.with_ctx(ctx).optimize_expr_in_num_ctx(&mut n.arg);
         }
     }
@@ -2500,8 +2489,6 @@ impl VisitMut for Optimizer<'_> {
             };
             n.visit_mut_children_with(&mut *self.with_ctx(ctx));
         }
-
-        self.optimize_expr_in_bool_ctx(&mut n.test);
     }
 
     fn visit_mut_yield_expr(&mut self, n: &mut YieldExpr) {
