+++ ecmascript/minifier/src/compress/optimize/ops.rs
@@ -1,8 +1,6 @@
 use super::Optimizer;
-use crate::compress::optimize::bools::is_ok_to_negate_for_cond;
-use crate::compress::optimize::bools::is_ok_to_negate_rhs;
 use crate::compress::optimize::is_pure_undefined;
-use crate::debug::dump;
+use crate::compress::util::negate;
 use crate::util::make_bool;
 use crate::util::ValueExt;
 use std::mem::swap;
@@ -274,168 +272,10 @@ impl Optimizer<'_> {
         self.negate(e);
     }
 
-    /// Creates `!e` where e is the expression passed as an argument.
-    ///
-    /// # Note
-    ///
-    /// This method returns `!e` if `!!e` is given as a argument.
-    ///
-    /// TODO: Handle special cases like !1 or !0
+    /// See [negate] for details.
     pub(super) fn negate(&mut self, e: &mut Expr) {
-        let start_str = dump(&*e);
-
         self.changed = true;
-        match e {
-            Expr::Bin(bin @ BinExpr { op: op!("=="), .. })
-            | Expr::Bin(bin @ BinExpr { op: op!("!="), .. })
-            | Expr::Bin(bin @ BinExpr { op: op!("==="), .. })
-            | Expr::Bin(bin @ BinExpr { op: op!("!=="), .. }) => {
-                bin.op = match bin.op {
-                    op!("==") => {
-                        op!("!=")
-                    }
-                    op!("!=") => {
-                        op!("==")
-                    }
-                    op!("===") => {
-                        op!("!==")
-                    }
-                    op!("!==") => {
-                        op!("===")
-                    }
-                    _ => {
-                        unreachable!()
-                    }
-                };
-                self.changed = true;
-                log::debug!("negate: binary");
-                return;
-            }
-
-            Expr::Bin(BinExpr {
-                left,
-                right,
-                op: op @ op!("&&"),
-                ..
-            }) if is_ok_to_negate_rhs(&right) => {
-                log::debug!("negate: a && b => !a || !b");
-
-                self.negate(&mut **left);
-                self.negate(&mut **right);
-                *op = op!("||");
-                return;
-            }
-
-            Expr::Bin(BinExpr {
-                left,
-                right,
-                op: op @ op!("||"),
-                ..
-            }) if is_ok_to_negate_rhs(&right) => {
-                log::debug!("negate: a || b => !a && !b");
-
-                self.negate(&mut **left);
-                self.negate(&mut **right);
-                *op = op!("&&");
-                return;
-            }
-
-            Expr::Cond(CondExpr { cons, alt, .. })
-                if is_ok_to_negate_for_cond(&cons) && is_ok_to_negate_for_cond(&alt) =>
-            {
-                log::debug!("negate: cond");
-
-                self.negate(&mut **cons);
-                self.negate(&mut **alt);
-                return;
-            }
-
-            Expr::Seq(SeqExpr { exprs, .. }) => {
-                if let Some(last) = exprs.last_mut() {
-                    log::debug!("negate: seq");
-
-                    self.negate(&mut **last);
-                    return;
-                }
-            }
-
-            _ => {}
-        }
-
-        let mut arg = Box::new(e.take());
-
-        match &mut *arg {
-            Expr::Unary(UnaryExpr {
-                op: op!("!"), arg, ..
-            }) => match &mut **arg {
-                Expr::Unary(UnaryExpr { op: op!("!"), .. }) => {
-                    log::debug!("negate: !!bool => !bool");
-                    *e = *arg.take();
-                    return;
-                }
-                Expr::Bin(BinExpr { op: op!("in"), .. })
-                | Expr::Bin(BinExpr {
-                    op: op!("instanceof"),
-                    ..
-                }) => {
-                    log::debug!("negate: !bool => bool");
-                    *e = *arg.take();
-                    return;
-                }
-                _ => {
-                    if self.ctx.in_bool_ctx {
-                        log::debug!("negate: !expr => expr (in bool context)");
-                        *e = *arg.take();
-                        return;
-                    }
-                }
-            },
-
-            _ => {}
-        }
-
-        log::debug!("negate: e => !e");
-
-        *e = Expr::Unary(UnaryExpr {
-            span: DUMMY_SP,
-            op: op!("!"),
-            arg,
-        });
-
-        if cfg!(feature = "debug") {
-            log::trace!("[Change] Negated `{}` as `{}`", start_str, dump(&*e));
-        }
-    }
-
-    pub(super) fn handle_negated_seq(&mut self, n: &mut Expr) {
-        match &mut *n {
-            Expr::Unary(e @ UnaryExpr { op: op!("!"), .. })
-            | Expr::Unary(
-                e @ UnaryExpr {
-                    op: op!("delete"), ..
-                },
-            ) => {
-                match &mut *e.arg {
-                    Expr::Seq(SeqExpr { exprs, .. }) => {
-                        if exprs.is_empty() {
-                            return;
-                        }
-                        log::debug!("optimizing negated sequences");
-
-                        {
-                            let last = exprs.last_mut().unwrap();
-                            self.optimize_expr_in_bool_ctx(last);
-                            // Negate last element.
-                            self.negate(last);
-                        }
-
-                        *n = *e.arg.take();
-                    }
-                    _ => {}
-                }
-            }
-            _ => {}
-        }
+        negate(e, self.ctx.in_bool_ctx)
     }
 
     /// This method does
